Move tests for moved code to the correct test class
Move tests for buildBinaryOperationTable to the relevant test class now that it is moved to WorkspaceOpOverloads.
Modify the tests to use a WorkspaceTester helper class rather than a real EventWorkspace, which is not available in the API module.

Re #23012
